Reorganized madrigal helper scripts.
