+++ mivit/helper/__init__.py
@@ -1,6 +1,7 @@
 import mango
 import superdarn
 import amisr
+import madrigal
 # from SuperDARN_dataset import *
 # from Madrigal_dataset import *
 # from AMISR_dataset import *
